requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1705494584803/work
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
anndata==0.9.2
astropy @ file:///home/conda/feedstock_root/build_artifacts/astropy_1667514593348/work
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
biopython==1.83
biotite==0.34.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
cell2cell==0.7.4
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
comm==0.2.2
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.8
decorator==5.1.1
dill==0.4.0
einops==0.8.1
einx==0.3.0
et_xmlfile==2.0.0
executing==2.1.0
fast-histogram @ file:///tmp/build/80754af9/fast-histogram_1614079150684/work
filelock @ file:///croot/filelock_1700591183607/work
-e git+https://github.com/microsoft/foldingdiff@18c865338db937fbd2ffba1e4fdfbcfdb99de8b7#egg=foldingdiff
fonttools @ file:///croot/fonttools_1713551344105/work
frozendict==2.4.6
frozenlist==1.5.0
fsspec @ file:///croot/fsspec_1724855583730/work
future @ file:///home/conda/feedstock_root/build_artifacts/future_1708610096684/work
get-annotations==0.1.2
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1711991025291/work
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1715527302982/work
google-auth==2.40.1
google-auth-oauthlib==1.0.0
grpcio @ file:///croot/grpc-split_1716834572340/work
gseapy==1.0.3
h5py==3.11.0
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1733636895686/work
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
ipykernel==6.29.5
jaxtyping==0.2.19
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
joblib @ file:///croot/joblib_1718217211762/work
jupyter_core==5.7.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
kneed==0.8.5
lightning-utilities @ file:///home/conda/feedstock_root/build_artifacts/lightning-utilities_1729059767967/work
llvmlite==0.41.1
loompy==3.0.8
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899923320/work
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1688684855958/work
matplotlib-inline==0.1.7
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpl-scatter-density @ file:///home/conda/feedstock_root/build_artifacts/mpl-scatter-density_1745585533847/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
msgpack @ file:///opt/conda/conda-bld/msgpack-python_1652362659880/work
multidict==6.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
numba==0.58.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682952083030/work
numpy-groupies==0.9.22
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
OmegaFold==0.0.0
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
parso==0.8.4
patsy==1.0.1
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.1
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyDeprecate @ file:///home/conda/feedstock_root/build_artifacts/pydeprecate_1623410077719/work
pyerfa @ file:///tmp/build/80754af9/pyerfa_1621560806183/work
Pygments==2.18.0
PyMuPDF==1.24.11
pynndescent==0.5.13
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyPDF2==3.0.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytorch-lightning @ file:///home/conda/feedstock_root/build_artifacts/pytorch-lightning_1658040017668/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1726669635272/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
requests-oauthlib==2.0.0
rsa==4.9.1
sacremoses @ file:///tmp/build/80754af9/sacremoses_1633107328213/work
scanpy==1.9.8
scikit-learn==1.3.2
scipy==1.10.1
scvelo==0.3.3
seaborn==0.13.2
session_info==1.0.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
stack-data==0.6.3
statannotations==0.7.2
statsmodels==0.14.1
stdlib-list==0.10.0
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1728484478345/work
tensorboard==2.14.0
tensorboard-data-server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1695425377577/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=11751b28e44b2a8840545cbda78535d37ed70e6cc0838e075f96451c4110e79a
tensorly==0.8.1
threadpoolctl==3.5.0
tokenizers @ file:///tmp/build/80754af9/tokenizers_1639607818835/work
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics @ file:///home/conda/feedstock_root/build_artifacts/torchmetrics_1731096877504/work
torchvision==0.13.1+cu113
tornado==6.4.1
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets==5.14.3
transformers==4.11.3
triton @ file:///opt/conda/envs/bld/conda-bld/torchtriton_1721056825793/work/python
typeguard==4.4.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2025.2
umap-learn==0.5.7
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1729884156362/work
xlrd==2.0.2
yarl==1.15.2
zipp @ file:///croot/zipp_1729012354496/work
